v2.2.8.7 - Possible fix for regional Inkbunny issue..
I had an issue where direct uploading to e621 through reBot resulted in an error, but it was on the e621 side.

My area, Brazil, uses inkbunny links with a "br" prefix, and they me numbered or not.

Inkbunny can redirect users from "br" to "brX" when needed, but according to Kira on discord, they don't want e621 to follow such redirects.

So I created this function to fix the link before it reaches the e621 image fetcher, so it always uses the already redirected link, in a certain way.

Created it as a function in case there are more regional links with this issue.