v2.2.8.8 - Fixed twitter not finding media in Brazil
In my region, Brazil, twitter translates one specific part of the html from "Image" to "imagem", resulting in ImageNodes returning Null.

I added a condition checking if ImageNodes is null, and if true, I tell it to try to assign again to the translated alternative.

Should other regions have this issue, it should be easier to add those to the condition I made.